+++ shaders/src/light_directional.fs
@@ -38,18 +38,19 @@ void evaluateDirectionalLight(const MaterialInputs material,
     float visibility = 1.0;
 #if defined(HAS_SHADOWING)
     if (light.NoL > 0.0) {
-
-        float ssContactShadowOcclusion = screenSpaceContactShadow(light.l);
-
-        if (frameUniforms.directionalShadows) {
+        float ssContactShadowOcclusion = 0.0;
+        if ((frameUniforms.directionalShadows & 1u) != 0u) {
+            if (frameUniforms.ssContactShadowDistance != 0.0) {
+                if (objectUniforms.screenSpaceContactShadows != 0) {
+                    ssContactShadowOcclusion = screenSpaceContactShadow(light.l);
+                }
+            }
             visibility = shadow(light_shadowMap, 0u, getLightSpacePosition());
             #if defined(MATERIAL_HAS_AMBIENT_OCCLUSION)
             visibility *= computeMicroShadowing(light.NoL, material.ambientOcclusion);
             #endif
         }
-
         visibility = visibility * (1.0 - ssContactShadowOcclusion);
-
     } else {
 #if defined(MATERIAL_CAN_SKIP_LIGHTING)
         return;